+++ src/aioverse/__init__.py
@@ -1,13 +1,9 @@
-# 服务模块
-from . import OpenAI
-
 # 基础模块 协议接口
 from . import errors
 from . import protocols
 
 
 __all__ = [
-	"OpenAI",
 	"errors",
 	"protocols",
 ]