+++ src/aioverse/core/__init__.py
@@ -1,0 +1,14 @@
+from ._transport			import iter_stream_chunks
+from .chat_completion		import chat_completion
+from .sse					import DONE, MAX_SSE_PARSE_ERRORS, READ_CHUNK_SIZE, SSEDecoder, StreamChunkParser
+
+
+__all__ = [
+	"chat_completion",
+	"iter_stream_chunks",
+	"DONE",
+	"MAX_SSE_PARSE_ERRORS",
+	"READ_CHUNK_SIZE",
+	"SSEDecoder",
+	"StreamChunkParser",
+]